+++ domain/stats.ts
