+++ entrypoints/content.ts
@@ -9,8 +9,8 @@ import {
   setupLeetcodeAutoReset,
   Tooltip,
 } from '@/content';
+import type { ProblemDescriptor } from '@/domain/cards';
 import { getServiceTranslations } from '@/services/i18n';
-import type { ProblemDescriptor } from '@/shared/cards';
 import type { Translations } from '@/shared/i18n';
 import { sendMessage } from '@/shared/messages';
