+++ entrypoints/popup/hooks/__tests__/useNoteEditor.test.tsx
@@ -1,9 +1,9 @@
 /** @vitest-environment happy-dom */
 import { act, renderHook, waitFor } from '@testing-library/react';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
+import { NOTES_MAX_LENGTH, type Note } from '@/domain/notes';
 import { noteQueryKeys } from '@/entrypoints/popup/queries/notes';
 import { sendMessage } from '@/shared/messages';
-import { NOTES_MAX_LENGTH, type Note } from '@/shared/notes';
 import { createDeferred } from '@/test/utils/deferred';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
