+++ entrypoints/popup/views/home/NotesSection.tsx
@@ -1,8 +1,8 @@
 import { useState } from 'react';
 import { Button, Label, TextArea, TextField } from 'react-aria-components';
+import { NOTES_MAX_LENGTH } from '@/domain/notes';
 import { useNoteEditor } from '@/entrypoints/popup/hooks/useNoteEditor';
 import { bounceButton } from '@/entrypoints/popup/styles';
-import { NOTES_MAX_LENGTH } from '@/shared/notes';
 import { useI18n } from '../../contexts/I18nContext';
 
 interface NotesSectionProps {
