+++ entrypoints/popup/views/home/ReviewQueue.tsx
@@ -1,5 +1,6 @@
 import { useState } from 'react';
 import type { Grade } from 'ts-fsrs';
+import type { Card, RateCardInput } from '@/domain/cards';
 import {
   useDelayCardMutation,
   usePauseCardMutation,
@@ -7,7 +8,6 @@ import {
   useRemoveCardMutation,
   useReviewQueueQuery,
 } from '@/entrypoints/popup/queries/cards';
-import type { Card, RateCardInput } from '@/shared/cards';
 import { useI18n } from '../../contexts/I18nContext';
 import { ActionsSection } from './ActionsSection';
 import { NotesSection } from './NotesSection';
