+++ entrypoints/popup/views/stats/__tests__/ReviewHistoryChart.test.tsx
@@ -5,9 +5,9 @@
 import { render, screen } from '@testing-library/react';
 import { Rating } from 'ts-fsrs';
 import { describe, expect, it, vi } from 'vitest';
+import type { DailyStats } from '@/domain/stats';
 import { statsQueryKeys } from '@/entrypoints/popup/queries/stats';
 import { sendMessage } from '@/shared/messages';
-import type { DailyStats } from '@/shared/stats';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
 import { ReviewHistoryChart } from '../ReviewHistoryChart';
