+++ entrypoints/popup/views/stats/__tests__/UpcomingReviewsChart.test.tsx
@@ -4,9 +4,9 @@
 
 import { render, screen } from '@testing-library/react';
 import { describe, expect, it, vi } from 'vitest';
+import type { UpcomingReviewStats } from '@/domain/stats';
 import { statsQueryKeys } from '@/entrypoints/popup/queries/stats';
 import { sendMessage } from '@/shared/messages';
-import type { UpcomingReviewStats } from '@/shared/stats';
 import { createDeferred } from '@/test/utils/deferred';
 import { createMessageMock } from '@/test/utils/message-mocks';
 import { createTestWrapper } from '@/test/utils/test-wrapper';
