+++ services/__tests__/import-export.test.ts
@@ -2,8 +2,8 @@ import { createEmptyCard, Rating } from 'ts-fsrs';
 import { afterEach, beforeEach, describe, expect, it } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { storage } from 'wxt/utils/storage';
-import type { Note } from '@/shared/notes';
-import type { DailyStats } from '@/shared/stats';
+import type { Note } from '@/domain/notes';
+import type { DailyStats } from '@/domain/stats';
 import { buildSettings } from '@/test/utils/settings-mocks';
 import type { StoredCard } from '../cards';
 import { applyImportData, exportData, importData, prepareImportData, resetAllData } from '../import-export';
