+++ services/__tests__/notes.test.ts
@@ -1,7 +1,7 @@
 import { beforeEach, describe, expect, it } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { storage } from 'wxt/utils/storage';
-import { NOTES_MAX_LENGTH, type Note } from '@/shared/notes';
+import { NOTES_MAX_LENGTH, type Note } from '@/domain/notes';
 import { deleteNote, getNote, saveNote } from '../notes';
 import { getNoteStorageKey } from '../storage-keys';
