+++ services/__tests__/stats.test.ts
@@ -2,8 +2,8 @@ import { State as FsrsState, Rating } from 'ts-fsrs';
 import { afterEach, beforeEach, describe, expect, it, vi } from 'vitest';
 import { fakeBrowser } from 'wxt/testing/fake-browser';
 import { storage } from 'wxt/utils/storage';
-import type { Difficulty } from '@/shared/cards';
-import type { DailyStats } from '@/shared/stats';
+import type { Difficulty } from '@/domain/cards';
+import type { DailyStats } from '@/domain/stats';
 import type { StoredCard } from '../cards';
 import { addCard } from '../cards';
 import {
