+++ services/i18n.ts
@@ -1,6 +1,6 @@
 import { storage } from '#imports';
+import type { Language } from '@/domain/settings';
 import { type Translations, translations } from '@/shared/i18n';
-import type { Language } from '@/shared/settings';
 import { STORAGE_KEYS } from './storage-keys';
 
 const DEFAULT_LANGUAGE: Language = 'en';
