+++ test/utils/card-mocks.ts
@@ -1,5 +1,5 @@
 import type { State } from 'ts-fsrs';
-import type { Card, ProblemDescriptor } from '@/shared/cards';
+import type { Card, ProblemDescriptor } from '@/domain/cards';
 
 export function buildProblem(overrides: Partial<ProblemDescriptor> = {}): ProblemDescriptor {
   return {
